Add docs and restructure directories
