+++ docs/source/guides.md
@@ -1,0 +1,24 @@
+# Getting Started
+
+IAI-Powered Carla SimaaS:
+Inverted-AI client [package](https://pypi.org/project/iai-client/)
+
+## Models
+
+## Start with an instruction
+
+## Add some examples
+
+## Define your settings
+
+## Train your models
+
+### Setup
+
+### API
+
+### Run the simulation
+
+### Understand the code
+
+### Closing
