+++ docs/source/index.md
@@ -1,0 +1,38 @@
+---
+hide-toc: true
+---
+
+```{include} overview.md
+:relative-images:
+```
+
+```{toctree}
+:hidden:
+:maxdepth: 4
+
+overview
+quickstart
+```
+
+```{toctree}
+:caption: GUIDES
+:hidden:
+:maxdepth: 2
+
+
+guides
+examples
+apireference
+miscellaneous
+modules/modules
+```
+
+```{eval-rst}
+
+REFERENCES
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
+```
