+++ docs/source/modules/invertedai_carlasim.rst
@@ -1,0 +1,21 @@
+Simulator packages
+====================================
+
+Submodules
+----------
+
+carla_examples.utils module
+--------------------------------------
+
+.. automodule:: simulators.carla_simulator.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: simulators.carla_simulator
+   :members:
+   :undoc-members:
+   :show-inheritance:
