+++ docs/source/modules/invertedai_drive.rst
@@ -1,0 +1,21 @@
+invertedai\_drive package
+============================
+
+Submodules
+----------
+
+invertedai\_drive.drive module
+-------------------------------------
+
+.. automodule:: invertedai_drive.drive
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: invertedai_drive
+   :members:
+   :undoc-members:
+   :show-inheritance:
