+++ docs/source/modules/modules.rst
@@ -1,0 +1,8 @@
+Inverted AI Drive
+===================
+
+.. toctree::
+   :maxdepth: 4
+
+   invertedai_drive
+   invertedai_carlasim
