+++ docs/source/overview.md
@@ -1,0 +1,41 @@
+```{eval-rst}
+.. image:: ../images/banner.svg
+  :width: 600
+  :alt: Alternative text
+```
+
+# Overview
+
+<!-- Content -->
+
+Breaking Down Barriers to Level 4-5 Autonomy
+
+```{button-ref} quickstart
+:ref-type: doc
+:color: primary
+:class: sd-rounded-pill
+
+Get Started
+```
+
+- **Safety Validation**
+- **Synthetic Data Generation**
+- **Training and Testing**
+- **Real-World Results From Simulation**
+
+  ![](../images/top_camera.gif)
+
+Inverted-AI provides cutting-edge solutions for closing the SIM2Real gap for autonomous vehicle (AV) industry.
+Industry leading AV simulators such as [Carla](https://carla.org/) and [NVIDIA DRIVE Sim](https://www.nvidia.com/en-us/self-driving-cars/simulation/) have closed the SIM2Real gap and allow for obtaining real-world results from simulation by:
+
+- Accurate Physics Simulation
+- Photo-Realistic Scenes
+- Accurate Sensor Data
+
+However, one crucial piece of the puzzle is missing: **Human-Like NPCs** (non player characters).
+State-of-the-art probabilistic behavioral models of Inverted-AI quantitatively and qualitatively outperform exiting logic-based and data-driven NPCs.
+
+Inverted-AI also provides another product:
+
+- Simulation as a service(SimaaS): AV simulation + Inverted-AI NPCs
+  - Currently [Carla](https://carla.org/) ([Huawei SMARTS](https://github.com/huawei-noah/SMARTS) and [NVIDIA DRIVE Sim](https://www.nvidia.com/en-us/self-driving-cars/simulation/) in production.)
