+++ docs/source/quickstart.md
@@ -1,0 +1,27 @@
+# Get Started
+
+## Installation
+
+[pypi-badge]: https://badge.fury.io/py/invertedai_drive.svg
+[pypi-link]: https://pypi.org/project/invertedai-drive/
+
+To install use [![PyPI][pypi-badge]][pypi-link]:
+
+```bash
+pip install invertedai_drive
+```
+
+## Running demo locally
+
+Download the examples [directory](https://github.com/inverted-ai/invertedai-drive/blob/master/examples) and run:
+
+```
+python -m venv .venv
+source .venv/bin/activate
+pip install -r requirements.txt
+.venv/bin/jupyter notebook Drive-Demo.ipynb
+```
+
+## Running demo in Colab
+
+[![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/inverted-ai/invertedai-drive/blob/develop/examples/Colab-Demo.ipynb)
