+++ invertedai_drive/drive.py
@@ -25,6 +25,26 @@ class Config:
 
 
 class Drive:
+    """
+    A class to represent a person.
+
+    ...
+
+    Attributes
+    ----------
+    name : str
+        first name of the person
+    surname : str
+        family name of the person
+    age : int
+        age of the person
+
+    Methods
+    -------
+    info(additional=""):
+        Prints the person's name and age.
+    """
+
     def __init__(self, config) -> None:
         self.location = config.location
         self.config = config
