+++ pyproject.toml
@@ -10,13 +10,33 @@ packages = [
 ]
 
 [tool.poetry.dependencies]
-python = ">=3.7.1,<3.11"
+python = ">=3.8,<3.11"
 torch = "^1.11.0"
 python-dotenv = "^0.20.0"
 numpy = "1.21.2"
+ipython = "^8.4.0"
+ipywidgets = "^8.0.1"
+matplotlib = "^3.5.3"
 
 [tool.poetry.dev-dependencies]
 twine = "==4.0.0"
+Sphinx = "^5.1.1"
+myst-parser = "^0.18.0"
+furo = "^2022.6.21"
+sphinx-rtd-theme = "^1.0.0"
+sphinx-copybutton = "^0.5.0"
+sphinx_design = "^0.3.0"
+sphinxcontrib-applehelp = "^1.0.2"
+sphinxcontrib-devhelp = "^1.0.2"
+sphinxcontrib-htmlhelp = "^2.0.0"
+sphinxcontrib-jsmath = "^1.0.1"
+sphinxcontrib-qthelp = "^1.0.3"
+sphinxcontrib-serializinghtml = "^1.1.5"
+sphinx-autobuild = "^2021.3.14"
+nbsphinx = "^0.8.9"
+jupyterlab = "^3.4.5"
+pydata-sphinx-theme = "^0.10.1"
+carla = "^0.9.13"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
