+++ simulators/carla_simulator/docker-compose.yml
@@ -1,0 +1,18 @@
+version: "3.3"
+
+services:
+  server:
+    image: "carlasim/carla"
+    network_mode: "host"
+    # restart: always
+    environment:
+      DISPLAY: $DISPLAY
+    deploy:
+      resources:
+        reservations:
+          devices:
+            - driver: nvidia
+              count: 1
+              capabilities: [ gpu ]
+    privileged: true
+    # command: ["./CarlaUE4.sh", "-fps=10"]
